Convert to workspace
Rather than using feature flags to build different binaries, we create
a cargo workspace with `crates/engine` containing the game logic, and
`crates/client` containing the user-facing stuff. In the future,
`crates/server` will contain the game server.

Note: We are currently getting a bevy panic after grenades explode. It
is under investigation, but I wanted to commit this so as to not lose
track of changes.