Restructure collisions and death
We add a `DeathCallback` to handle events on death, such as explosions.
It is likely that we will want to refactor this, as it currently
require an `enum` with all possible callbacks.

We also add a better `collision_system`, getting rid of the mess we had
before. Now, entities can easily opt-in to tracking collisions, and can
run their own systems for what to do when that happens.