Fix character movement
Our character colliders were cylinders. I'm not sure why, it might be
because we used to use a coordinate system with Z-up, and at the time it
was difficult to make a capsule in the orientation we wanted. Or maybe
another reason.

In any case, we can use capsules now, which makes rapier much happier.
They also have a `ContactSkin` which should improve things. I am unsure
why rapier got worse at handling cylinders.